Move `TLS` config options to `config#TLS`
