+++ internal/config/config.go
@@ -1,8 +1,6 @@
 package config
 
 import (
-	"crypto/tls"
-	"crypto/x509"
 	"github.com/creasty/defaults"
 	"github.com/goccy/go-yaml"
 	"github.com/icinga/icingadb/pkg/logging"
@@ -82,54 +80,4 @@ func ParseFlags() (*Flags, error) {
 	}
 
 	return f, nil
-}
-
-// TLS provides TLS configuration options for Redis and Database.
-type TLS struct {
-	Enable   bool   `yaml:"tls"`
-	Cert     string `yaml:"cert"`
-	Key      string `yaml:"key"`
-	Ca       string `yaml:"ca"`
-	Insecure bool   `yaml:"insecure"`
-}
-
-// MakeConfig assembles a tls.Config from t and serverName.
-func (t *TLS) MakeConfig(serverName string) (*tls.Config, error) {
-	if !t.Enable {
-		return nil, nil
-	}
-
-	tlsConfig := &tls.Config{}
-	if t.Cert == "" {
-		if t.Key != "" {
-			return nil, errors.New("private key given, but client certificate missing")
-		}
-	} else if t.Key == "" {
-		return nil, errors.New("client certificate given, but private key missing")
-	} else {
-		crt, err := tls.LoadX509KeyPair(t.Cert, t.Key)
-		if err != nil {
-			return nil, errors.Wrap(err, "can't load X.509 key pair")
-		}
-
-		tlsConfig.Certificates = []tls.Certificate{crt}
-	}
-
-	if t.Insecure {
-		tlsConfig.InsecureSkipVerify = true
-	} else if t.Ca != "" {
-		raw, err := os.ReadFile(t.Ca)
-		if err != nil {
-			return nil, errors.Wrap(err, "can't read CA file")
-		}
-
-		tlsConfig.RootCAs = x509.NewCertPool()
-		if !tlsConfig.RootCAs.AppendCertsFromPEM(raw) {
-			return nil, errors.New("can't parse CA file")
-		}
-	}
-
-	tlsConfig.ServerName = serverName
-
-	return tlsConfig, nil
 }
